Two crashes, the admin bypass, four UI bugs, and seeding for the new tables
## Crashes on dev

`weeks` arrives from a form as the string "4" and Carbon's addWeeks() declares
int|float, so creating any availability poll was a TypeError. Validation said
`integer`, which checks the value is numeric and does not change its type --
the trap is that it reads like a cast and is not one.

The exam page died on a null poll. The real cause was the policy.

## ExamPolicy::before() bypassed the whole workflow

It returned true for an admin on EVERY ability, which short-circuits the stage
checks below it -- so an admin was offered the student's availability form on
an exam nobody had authorised, and the page then died on a poll that does not
exist until authorisation creates it.

The stage ordering IS the feature. An admin skipping it is not a superpower,
it is the workflow failing to apply to the person most likely to be clicking
around it. before() is now limited to view, viewAny and cancel -- reading and
stopping, the two things genuinely needed at 2am when the examiner who took it
has vanished. Everything else falls through to its own check, which an admin
passes on permission anyway, at the right stage.

## The pages join the rest of Control Center

The automation log and both exam pages were Tailwind on a parallel layout with
their own sidebar, which made them the one place that looked like a different
application. That was right while the Tailwind layout was the destination and
wrong the moment the restyle became the destination instead.

They are Bootstrap now, on layouts.app, and they inherit the restyle for free
-- including anything it gains later. A page on a parallel layout is a second
thing to keep in step, for ever, for no benefit.

## Four UI bugs, all found by looking

A card header with `p-0` still got the chrome section's padding, so the orange
highlight on the reports archive sat between two teal strips instead of filling
the row. `p-0` is Bootstrap's own way of saying this header has no padding;
honour it rather than overriding and compensating.

Cards inside cards were being drawn as cards -- an accordion is a stack of
.card inside a .card-body, and every one had a radius, a border and a shadow.
Nested cards are rows.

The topbar search is `.bg-light.border-0`, and --bg-light in dark mode is the
card-header teal, on a topbar that is nearly the same colour. A search box you
cannot see is one nobody uses.

`.form-select` is width:100%, so in a flex card header it shrank until "All
categories" became "All categori" with the chevron on top of it. Sized to
content with a floor, and told not to shrink.

Plus spacing between stacked alerts, which is what made the green and blue
blocks on the dashboard read as one two-tone box.

## The desk move is a modal

The inline dropdown showed the current desk twice -- once as the badge above it
and once as its own selected option -- made every row two lines tall for
something used weekly, and read as an edit control, which a select that changes
nothing until you press a button beside it always does.

## Availability and exams move to Training

They sat beside Booking, putting "My availability" next to a bookings calendar
it has nothing to do with. Ungated, deliberately: a student needs to find their
own exam and give their own availability, and holds no training permission at
all.

## Seeding

One exam at every stage of the workflow, because each shows a different page
and seeding one exercises one. The awkward states are seeded on purpose: an
exam inside the seven-day window with paperwork outstanding, and one whose
cleared slots have aged out. Those are the two the daily sweep exists to catch
and neither can be reached by clicking the happy path.

Plus availability responses with a real subset intersection rather than the
student's whole list, and an automation log with both kinds of row -- warnings
especially, since the page defaults to them and one that opens empty on dev is
one nobody looks at twice.

+++ app/Http/Controllers/Vatssa/AvailabilityController.php
@@ -99,13 +99,17 @@ class AvailabilityController extends Controller
             'purpose' => $data['purpose'],
             'description' => $data['description'] ?? null,
             'starts_on' => $from,
-            'ends_on' => $from->addWeeks($data['weeks'] ?? 4)->subDay(),
+            // (int), because form input is a STRING. Carbon's addWeeks()
+            // declares int|float and throws a TypeError on "4" -- validation
+            // said `integer`, which checks the value is numeric and does not
+            // change its type.
+            'ends_on' => $from->addWeeks((int) ($data['weeks'] ?? 4))->subDay(),
             'training_id' => $data['training_id'] ?? null,
             'created_by' => Auth::id(),
             // Half an hour for a session or an exam; an hour for a meeting,
             // where nobody schedules to the half hour and the finer grid is
             // just more cells to drag across.
-            'slot_minutes' => $data['purpose'] === AvailabilityPoll::MEETING ? 60 : 30,
+            'slot_minutes' => (int) ($data['purpose'] === AvailabilityPoll::MEETING ? 60 : 30),
         ]);
 
         return redirect()->route('vatssa.availability.show', $poll)
